Rename lib and Pics
